+++ CHANGELOG.rst
@@ -10,6 +10,7 @@
 * optionally specify puppet packages to install (e.g. to pin a version)
 * Order scripts by numbers instead of characters (#2)
 * provide provison step saltstack_provision
+* Manage repository sources (#4)
 
 v1.2.0 (2015-04-30)
 -------------------